agent/requirements.txt
mcp
python-json-logger
openai
fastapi
uvicorn[standard]
requests
pydantic
black
PyJWT
